Prevent duplicate callback on fail
I was trying to figure out why after testing a file for expected failure my callback function was getting mysteriously triggered again later on.  I looked here in the source for a clue and noticed the fs.read call in the exported compare function triggers regardless of whether the initial file size check is off or not.  Please commit this correction.  Thanks!